refactor: create node.py
