+++ tests/unit_tests/test_line.py
@@ -3,8 +3,8 @@ from typing import List
 
 import pytest
 
-from sqlfmt.exception import SqlfmtBracketError
-from sqlfmt.line import Comment, Line, Node
+from sqlfmt.comment import Comment
+from sqlfmt.line import Line
 from sqlfmt.mode import Mode
 from sqlfmt.token import Token, TokenType
 from tests.util import read_test_data
@@ -48,48 +48,6 @@ def simple_line(tokens: List[Token], bare_line: Line) -> Line:
     for token in tokens:
         simple_line.append_token(token)
     return simple_line
-
-
-def test_calculate_depth() -> None:
-    select_t = Token(
-        type=TokenType.UNTERM_KEYWORD,
-        prefix="",
-        token="select",
-        spos=0,
-        epos=6,
-    )
-    select_n = Node.from_token(token=select_t, previous_node=None)
-    assert (select_n.depth, select_n.open_brackets) == ((0, 0), [])
-
-    open_paren_t = Token(
-        type=TokenType.BRACKET_OPEN,
-        prefix="    ",
-        token="(",
-        spos=8,
-        epos=9,
-    )
-    open_paren_n = Node.from_token(token=open_paren_t, previous_node=select_n)
-    assert (open_paren_n.depth, open_paren_n.open_brackets) == ((1, 0), [select_n])
-
-    one_t = Token(
-        type=TokenType.NUMBER,
-        prefix=" ",
-        token="1",
-        spos=10,
-        epos=11,
-    )
-    one_n = Node.from_token(token=one_t, previous_node=open_paren_n)
-    assert (one_n.depth, one_n.open_brackets) == ((2, 0), [select_n, open_paren_n])
-
-    close_paren_t = Token(
-        type=TokenType.BRACKET_CLOSE,
-        prefix="",
-        token=")",
-        spos=11,
-        epos=12,
-    )
-    close_paren_n = Node.from_token(token=close_paren_t, previous_node=one_n)
-    assert (close_paren_n.depth, close_paren_n.open_brackets) == ((1, 0), [select_n])
 
 
 def test_bare_line(bare_line: Line) -> None:
@@ -363,20 +321,6 @@ def test_is_standalone_jinja_statement(bare_line: Line, simple_line: Line) -> No
     assert not simple_line.is_standalone_jinja_statement
 
 
-def test_calculate_depth_exception() -> None:
-
-    close_paren = Token(
-        type=TokenType.BRACKET_CLOSE,
-        prefix="",
-        token=")",
-        spos=0,
-        epos=1,
-    )
-
-    with pytest.raises(SqlfmtBracketError):
-        _ = Node.from_token(close_paren, previous_node=None)
-
-
 def test_closes_bracket_from_previous_line(
     simple_line: Line, default_mode: Mode
 ) -> None:
@@ -494,98 +438,3 @@ def test_is_multiplication_star_bare_line(bare_line: Line) -> None:
     bare_line.append_token(star)
     assert bare_line.nodes[0].token == star
     assert not bare_line.nodes[0].is_multiplication_star
-
-
-def test_jinja_depth(default_mode: Mode) -> None:
-    source_string, _ = read_test_data("unit_tests/test_line/test_jinja_depth.sql")
-    q = default_mode.dialect.initialize_analyzer(
-        line_length=default_mode.line_length
-    ).parse_query(source_string=source_string)
-    expected = [
-        (0, 0),  # {{ config(materialized="table") }}
-        (0, 0),  # \n
-        (0, 0),  # \n
-        (0, 0),  # {%- set n = 5 -%}
-        (0, 0),  # \n
-        (0, 0),  # with
-        (1, 0),  # \n
-        (1, 0),  # {% for i in range(n) %}
-        (1, 1),  # \n
-        (1, 1),  # dont_do_this_
-        (1, 1),  # {{ i }}
-        (1, 1),  # as
-        (1, 1),  # (
-        (2, 1),  # \n
-        (2, 1),  # {% if foo %}
-        (2, 2),  # \n
-        (2, 2),  # select
-        (3, 2),  # \n
-        (2, 1),  # {% elif bar %}
-        (2, 2),  # \n
-        (2, 2),  # select distinct
-        (3, 2),  # \n
-        (2, 1),  # {% elif baz %}
-        (2, 2),  # \n
-        (2, 2),  # select top 25
-        (3, 2),  # \n
-        (2, 1),  # {% else %}
-        (2, 2),  # \n
-        (2, 2),  # select
-        (3, 2),  # \n
-        (3, 1),  # {% endif %}
-        (3, 1),  # \n
-        (3, 1),  # my_col
-        (3, 1),  # \n
-        (2, 1),  # from
-        (3, 1),  # \n
-        (3, 1),  # {% if i == qux %}
-        (3, 2),  # \n
-        (3, 2),  # zip
-        (3, 2),  # \n
-        (3, 1),  # {% else %}
-        (3, 2),  # \n
-        (3, 2),  # zap
-        (3, 2),  # \n
-        (3, 1),  # {% endif %}
-        (3, 1),  # \n
-        (1, 1),  # )
-        (1, 1),  # {% if not loop.last %}
-        (1, 2),  # ,
-        (1, 1),  # {% endif%}
-        (1, 1),  # \n
-        (1, 0),  # {% endfor %}
-        (1, 0),  # \n
-        (1, 0),  # {% for i in range(n) %}
-        (1, 1),  # \n
-        (0, 1),  # select
-        (1, 1),  # \n
-        (1, 1),  # *
-        (1, 1),  # \n
-        (0, 1),  # from
-        (1, 1),  # \n
-        (1, 1),  # dont_do_this_
-        (1, 1),  # {{ i }}
-        (1, 1),  # \n
-        (1, 1),  # {% if not loop.last -%}
-        (1, 2),  # \n
-        (0, 2),  # union all
-        (1, 2),  # \n
-        (0, 1),  # {%- endif %}
-        (0, 1),  # \n
-        (0, 0),  # {% endfor %}
-        (0, 0),  # \n
-    ]
-    actual = [node.depth for node in q.nodes]
-    assert actual == expected
-
-
-def test_from_token_raises_bracket_error_on_jinja_block_end() -> None:
-    t = Token(
-        type=TokenType.JINJA_BLOCK_END,
-        prefix="",
-        token="{% endif %}",
-        spos=0,
-        epos=11,
-    )
-    with pytest.raises(SqlfmtBracketError):
-        _ = Node.from_token(t, previous_node=None)
